requirements.txt
pywhatkit
mysql-connector-python
pymongo